requirements.txt
Flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
openai>=0.28.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
lxml>=4.9.0
selenium>=4.0.0
webdriver-manager>=4.0.0
schedule>=1.2.0
flask-cors>=4.0.0
